refactor(#20) :: routes 및 폴더 구조 변경
