+++ src/pages/register/terms/style.ts
