+++ src/routes/index.tsx
@@ -2,17 +2,20 @@ import { useRoutes } from 'react-router-dom';
 
 import Splash from '../pages/splash/';
 import Auth from 'pages/auth';
-import Terms from 'pages/terms';
+import Home from 'pages/home';
+import NotFound from 'pages/notfound';
+// register
+import Terms from 'pages/register/terms';
 import Info from 'pages/register/info';
-import SurveyEnd from '../pages/register/surveyEnd';
 import SurveyStyle from 'pages/register/surveyStyle';
 import SurveyFood from 'pages/register/surveyFood';
 import SurveyMBTI from 'pages/register/surveyMBTI';
-import Home from 'pages/home';
 import SurveryTMI from 'pages/register/surveyTMI';
-import Chat from 'pages/chat';
-import NotFound from 'pages/notfound';
-import SelectDate from 'pages/selectDate';
+import SurveyEnd from 'pages/register/surveyEnd';
+// plan
+import Chat from 'pages/plan/chat';
+import SelectDate from 'pages/plan/selectDate';
+// profile
 import Edit from 'pages/profile/edit';
 
 export default function Router() {
@@ -23,21 +26,26 @@ export default function Router() {
         { path: '/', element: <Home /> },
         { path: 'splash', element: <Splash /> },
         { path: 'auth', element: <Auth /> },
-        { path: 'terms', element: <Terms /> },
-        { path: 'chat', element: <Chat /> },
-        { path: 'selectdate', element: <SelectDate /> },
         { path: '*', element: <NotFound /> }
       ]
     },
     {
       path: '/register',
       children: [
+        { path: 'terms', element: <Terms /> },
         { path: 'info', element: <Info /> },
         { path: 'surveystyle', element: <SurveyStyle /> },
-        { path: 'surveyend', element: <SurveyEnd /> },
         { path: 'surveyfood', element: <SurveyFood /> },
         { path: 'surveymbti', element: <SurveyMBTI /> },
-        { path: 'surveytmi', element: <SurveryTMI /> }
+        { path: 'surveytmi', element: <SurveryTMI /> },
+        { path: 'surveyend', element: <SurveyEnd /> }
+      ]
+    },
+    {
+      path: 'plan',
+      children: [
+        { path: 'chat', element: <Chat /> },
+        { path: 'selectdate', element: <SelectDate /> }
       ]
     },
     {
